Refs #12161. Avoid including condition_variable to user
